BugFix: correct errors in createLabel()
The return value(s) from the Lua TLuaInterpreter sub-functions
* createLabelMainWindow(...)
* createLabelUserWindow(...)
were not being passed and handled correctly to the main createLabel()
function. Also the behaviour should a label be requested to be put on
a non-existent user-window was to place it on the main window rather
than to produce a run time error.

Also, there is no need to route the execution via the `mudlet` singleton
class which means that mdulet::create(...) can be removed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>