Refactor: remove confusing shadowing of cTelnet member with local variable
`(void) processTelnetCommand(const std::& command)` is confusing because
it uses as an argument a member's name - as it happens the member that gets
masked by this is the same argument as is processed inside the method.

It is less deceptive to have the method work on the member directly - and
as it is always followed by a line that clears the member (except for one
case where the member (a std::string) is reset to contain a single
character) - it is reasonable to move the clearing action to the end of the
method as well.

The member was `(std::string) cTelnet::command` but for clarity it has been
renamed to `cTelnet::mCommand`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>