W.I.P. Add un-decodable bytes as hidden hex digits into TBuffer data
The raw bytes are each converted to a pair of hex digits and those are
encoded as the unicode non-characters in the range U+FDD0 (`0`) to U+FDDF
(`F`) following on from the replacement character.

Further work is needed in order to show or hide the data as required; to
make the painting code convert them back to something that the user can
make sense of (or ignore) - and also tweak the text analyser to treat them
specially.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>