Improve: make minimum word length for spell-checking adjustable
Also modify it to use the number of graphemess ("visual text" units)
rather than the number of `QChar`s in the `QString` - this avoids
complications about measuring the "word length" when the text is encoded
in a *variable length encoding* (a `QString` is encoded in UTF-16BE form).

The setting is saved/loaded on a per-profile basis in the game save file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>