Infrastructure: add Qt6 Support
This is a work in progress because there is no support in the Qt6 API for
`QMediaPlaylist` - their recommendation is to import code from the Qt 5.15
examples in https://doc.qt.io/qt-6/qtmultimedia-player-example.html
with the code at:
https://code.qt.io/cgit/qt/qtmultimedia.git/tree/examples/multimedia/player?h=6.4
which has a copy of the code from the `QMediaPlayList` class.

Otherwise the code compiles fairly cleanly for Qt 6.4.1

Changes needed:
* `QMapIterator<T1, T2>::const_iterator` or `QMapIterator<T1, T1>` are
required to iterator over a `QMultiMap<T1, T2>` in Qt 5.x but
`QMultiMapIterator<T1, T2>::const_iterator` or ``QMultiMapIterator<T1, T2>`
must be used in Qt 6.x.
* `QTextStream` has lost the `setCodec(...)` method in Qt 6.x as encoding
conversion is now done by a `QStringConverter` class instead. Fortunately
in all the cases where we were forcing that class to use a UTF-8 encoding
(on Windows, where the Qt 5.x default was to use the system/local-8Bit
encoding) the default for Qt 6.x is to use the UTF-8 that we want.
* `QPait<T1, T2>` is no longer an explicit Qt class in Qt 6.x - instead it
is now just a typedef to `std::pair<T1, T2>` whilst this is not an issue
for most of our usages it is more awkward for our `MxpTagAttribute` which
is derived from it - which required some tweaking to get it to compile. I
have not completely removed all warnings about it's compilation - there
are still warnings about:
  * 'X' has virtual functions but non-virtual destructor [non-virtual-dtor]
  * 'X' has no out-of-line virtual method definitions; its vtable will be
emitted in every translation unit [weak-vtables]
* Qt 5.x classes that had `(int) size()` and `(int) count()` methods return
a new `qsizetype` in Qt 6.x which makes all the `static_cast`-ing needed
in classes like `TBuffer` when comparing `TChar` arrays to `QChar`
(`QString`) ones {the former used to hold the character attributes and the
latter the characters themselves} not required to shut up all the warnings
that the former differences in those types caused.
* `QMouseEvents` have lost `(QPoint) pos()` and `(QPoint) globalPos()`
methods, in order to get the same values it is necessary to use the
`(QPoinrF) position()` and `globalPosition()` methods and convert the
results to a `(QPoint)` with the `QPointF::toPoint()` method.
* `QMouseEvents` has renamed `localPos()` to `position()` in Qt 6.x.
* `QStringRef`s and other classes methods that return them have gone away
in Qt 6.x, instead there is a new `QStringView` class. In the few cases
where we were using them it was simplest to revert to the heavier `QString`
class for the moment.
* `QWidget::enterEvent(...)` requires a `QEnterEvent*` in Qt 6.x but a
`QEvent*` in Qt 5.x the error messages that this change provoked were quite
confusing when not handled in the `dlgTriggerEditor` and `TTextEdit`
classes.
* `Q_DISABLE_MOVE(...)` has been removed from Qt 6.x but where it is also
wanted alongside `Q_DISABLE_COPY(...)` there is a combined
`Q_DISABLE_COPY_MOVE(...)` that was introduced in Qt 5.13.
* We needed to generate some `QMouseEvents` in `TTextEdit` but the
constructor we were using has been deprecated/obsoleted in favour of two
others that take a "global" position as well as a "local" one, fortunately
the event we were generating got it's "local" position from a "global" one
anyhow, so it was simple to use one of the others for both Qt 5 and 6.
* The signature of `QtConcurrent::run(...)` has changed such that for our
usage the first two arguments have to be swapped around. Debugging that was
HARD. It was partly made so by the fact that the method to be run
`XMLexport::saveXml` was overloaded and working out how to resolve that
escaped me. So I renamed one of the methods `(QString) XMLexport::saveXml()`
to `saveXmlWorker()` which helped...
* `QXmlStreamReader::name()` now returns a `QStringView` in Qt 6.x and
those cannot be compared to a raw C-string literal with `operator==()`,
for simplicity it was easiest to convert all the comparisons in `XMLimport`
with an inverted `QStringView::compare(QLatin1String(...))` as that changes
back to a `QStringRef::compare(QLatin1String(...))` in the Qt 5,x case.
* `QAction::associatedWidgets()` is deprecated in Qt 6.x and instead one
must use `QAction::associatedObjects()` and find the ones that can be
`qobject_cast` to a `QWidget`.
* `QDateTime::fromTime_t(...)` was deprecated before Qt 5.14 - and has
been removed from Qt 6.x - the replacement is
`QDateTime::fromSecsSinceEpoch(...)` which is usable in our Qt 5.x builds.
* `QApplication::desktop()` was deprecated in Qt 5.15.2 and removed in
Qt 6.x - apart from anything else it is not that useful should the user
have multiple screens as it returns the whole desktop spread over those
screens and using its size to determine the sizing of windows (as we did
for the first time Mudlet was run) is wrong. Instead we can take something
about the size of the screen that the OS/DE opens Mudlet on from the screen
returned by `QApplication::primaryScreen()`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>