Fix: correct a code structuring error
Whilst running `lupdate` to update `translations/translated/mudlet_en_US.ts`
it reported that some translation metadata could not be parsed for the
'[ XXXX ] -` messages that the `cTelnet::postMessage(...)` processes.
Looking more closely I could see that `else if` was being used with the
result that some statement blocks would not be correctly executed. This
commit fixes them.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>